Fixing pylint erros, removing main eecution file and adding .pylintrc file with TODO information
